+++ TopQuarkAnalysis/TopKinFitter/interface/TtHadKinFitter.h
@@ -1,7 +1,7 @@
 #ifndef TtHadKinFitter_h
 #define TtHadKinFitter_h
 //
-// $Id: TtHadKinFitter.h,v 1.1 2007/10/06 15:13:07 mfhansen Exp $
+// $Id: TtHadKinFitter.h,v 1.2 2007/10/15 23:51:12 lowette Exp $
 // Adapted TtSemiKinFitter.h,v 1.3 2007/09/19 23:08:09 lowette Exp 
 // for fully hadronic channel 
 
@@ -13,57 +13,54 @@
 
 #include <vector>
 
-
 class TKinFitter;
 class TAbsFitParticle;
 class TFitConstraintM;
 
-
 class TtHadKinFitter {
 
-  public:
-
-    enum Parametrization { EMom, EtEtaPhi, EtThetaPhi };
-
-  public:
-
-    TtHadKinFitter();
-    TtHadKinFitter(int jetParam, int maxNrIter, double maxDeltaS, double maxF, std::vector<int> constraints);
-    TtHadKinFitter(Parametrization jetParam, int maxNrIter, double maxDeltaS, double maxF, std::vector<int> constraints);
-    ~TtHadKinFitter();
-
-    TtHadEvtSolution addKinFitInfo(TtHadEvtSolution * asol);
-
-  private:
-
-    void setupFitter();
-    std::vector<double> translateCovM(TMatrixD &);
-
-  private:
-
-    // the kinematic fitter
-    TKinFitter * theFitter_;
-    // the particles that enter the kinematic fit
-    TAbsFitParticle * fitHadb_;
-    TAbsFitParticle * fitHadp_;
-    TAbsFitParticle * fitHadq_;
-    TAbsFitParticle * fitHadbbar_;
-    TAbsFitParticle * fitHadj_;
-    TAbsFitParticle * fitHadk_;
-    // the constraints on the fit
-    TFitConstraintM  * cons1_;
-    TFitConstraintM  * cons2_;
-    TFitConstraintM  * cons3_;
-    TFitConstraintM  * cons4_;
-    TFitConstraintM  * cons5_;
-
-    // other parameters
-    Parametrization jetParam_;
-    int maxNrIter_;
-    double maxDeltaS_;
-    double maxF_;
-    std::vector<int> constraints_;
-
+ public:
+  
+  enum Parametrization { EMom, EtEtaPhi, EtThetaPhi };
+  
+ public:
+  
+  TtHadKinFitter();
+  TtHadKinFitter(int jetParam, int maxNrIter, double maxDeltaS, double maxF, std::vector<int> constraints);
+  TtHadKinFitter(Parametrization jetParam, int maxNrIter, double maxDeltaS, double maxF, std::vector<int> constraints);
+  ~TtHadKinFitter();
+  
+  TtHadEvtSolution addKinFitInfo(TtHadEvtSolution * asol);
+  
+ private:
+  
+  void setupFitter();
+  std::vector<float> translateCovM(TMatrixD &);
+  
+ private:
+  
+  // the kinematic fitter
+  TKinFitter * theFitter_;
+  // the particles that enter the kinematic fit
+  TAbsFitParticle * fitHadb_;
+  TAbsFitParticle * fitHadp_;
+  TAbsFitParticle * fitHadq_;
+  TAbsFitParticle * fitHadbbar_;
+  TAbsFitParticle * fitHadj_;
+  TAbsFitParticle * fitHadk_;
+  // the constraints on the fit
+  TFitConstraintM  * cons1_;
+  TFitConstraintM  * cons2_;
+  TFitConstraintM  * cons3_;
+  TFitConstraintM  * cons4_;
+  TFitConstraintM  * cons5_;
+  
+  // other parameters
+  Parametrization jetParam_;
+  int maxNrIter_;
+  double maxDeltaS_;
+  double maxF_;
+  std::vector<int> constraints_;  
 };
 
 #endif
